+++ tests/lib_2/js/static_interop_test/disallowed_generative_constructor_static_test.dart
@@ -1,11 +1,11 @@
-// Copyright (c) 2022, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
 // @dart = 2.9
 
 @JS()
-library generative_constructor_static_test;
+library disallowed_generative_constructor_static_test;
 
 import 'package:js/js.dart';